날씨를 검색 후 클릭 하였을 때 표시되는 NowFragment, OneWeekFragment가 표시되도록 코드 작성.

+++ app/src/main/java/com/lsb/lweather/models/nowWeather/Clouds.java
@@ -4,9 +4,9 @@ package com.lsb.lweather.models.nowWeather;
 
 public class Clouds {
 
-    private Integer all;
+    private double all;
 
-    public Integer getAll() {
+    public double getAll() {
         return all;
     }
